// project/src/types/database.types.ts
export interface Database {
  public: {
    Tables: {
      projects: {
        Row: {
          id: string
          created_at: string
          title: string
          description: string
          short_description: string | null
          tags: string[] | null
          image_url: string | null
          github_url: string | null
          live_url: string | null
          featured: boolean
          user_id: string
        }
        Insert: {
          id?: string
          created_at?: string
          title: string
          description: string
          short_description?: string | null
          tags?: string[] | null
          image_url?: string | null
          github_url?: string | null
          live_url?: string | null
          featured?: boolean
          user_id: string
        }
        Update: {
          id?: string
          created_at?: string
          title?: string
          description?: string
          short_description?: string | null
          tags?: string[] | null
          image_url?: string | null
          github_url?: string | null
          live_url?: string | null
          featured?: boolean
          user_id?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}